Reorder files for chaos mesh functionality
